requirements.txt
Django==2.0.2
django-filter==1.1.0
djangorestframework==3.7.7
Markdown==2.6.11
django-cors-headers==2.1.0
httpie==0.9.9